+++ ton_client/src/net/iterators/transaction_iterator/mod.rs
@@ -333,25 +333,25 @@ pub struct ParamsOfCreateTransactionIterator {
 ///
 /// Iterated item is a JSON objects with transaction data. The minimal set of returned
 /// fields is:
-///
+/// ```text
+/// id
+/// account_addr
+/// now
+/// balance_delta(format:DEC)
+/// bounce { bounce_type }
+/// in_message {
 ///     id
-///     account_addr
-///     now
-///     balance_delta(format:DEC)
-///     bounce { bounce_type }
-///     in_message {
-///         id
-///         value(format:DEC)
-///         msg_type
-///         src
-///     }
-///     out_messages {
-///         id
-///         value(format:DEC)
-///         msg_type
-///         dst
-///     }
-///
+///     value(format:DEC)
+///     msg_type
+///     src
+/// }
+/// out_messages {
+///     id
+///     value(format:DEC)
+///     msg_type
+///     dst
+/// }
+/// ```
 /// Application can request an additional fields in the `result` parameter.
 ///
 /// Another parameter that affects on the returned fields is the `include_transfers`.
@@ -363,7 +363,7 @@ pub struct ParamsOfCreateTransactionIterator {
 /// - isBounced – indicates that the transaction is bounced, which means the value will be returned back to the sender.
 /// - isDeposit – indicates that this transfer is the deposit (true) or withdraw (false).
 /// - counterparty – account address of the transfer source or destination depending on `isDeposit`.
-/// - value – amount of nano tokens transfered. The value is represented as a decimal string
+/// - value – amount of nano tokens transferred. The value is represented as a decimal string
 /// because the actual value can be more precise than the JSON number can represent. Application
 /// must use this string carefully – conversion to number can follow to loose of precision.
 ///
@@ -403,7 +403,7 @@ pub struct ParamsOfResumeTransactionIterator {
 
 /// Resumes transaction iterator.
 ///
-/// The iterator stays exactly at the same position where the `resume_state` was catched.
+/// The iterator stays exactly at the same position where the `resume_state` was caught.
 /// Note that `resume_state` doesn't store the account filter. If the application requires
 /// to use the same account filter as it was when the iterator was created then the application
 /// must pass the account filter again in `accounts_filter` parameter.